aggregates refactoring
